pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = 'arcgis-yelp'
version = "0.0.0"
authors = [
  { name="Joel McCune (https://github.com/knu2xs)"},
]
description = "Accessing the Yelp API in conjunction with ArcGIS"
readme = "README.md"
requires-python = ">=3.6, <=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
license = {file = "LICENSE"}
dependencies = [
    "azure-cosmos",
    "h3>=4.*",
    "requests",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["arcgis_yelp"]
